+++ client/task-list/task-list.js
@@ -8,13 +8,11 @@ import { useSelect, useDispatch } from '@wordpress/data';
 import {
 	EllipsisMenu,
 	Badge,
-	__experimentalList as List,
-	__experimentalCollapsibleList as CollapsibleList,
 } from '@woocommerce/components';
 import { updateQueryString } from '@woocommerce/navigation';
 import { OPTIONS_STORE_NAME, ONBOARDING_STORE_NAME } from '@woocommerce/data';
 import { recordEvent } from '@woocommerce/tracks';
-import { Text } from '@woocommerce/experimental';
+import { Text, List, CollapsibleList } from '@woocommerce/experimental';
 
 /**
  * Internal dependencies
