+++ packages/components/src/style.scss
@@ -18,7 +18,6 @@
 @import 'gravatar/style.scss';
 @import 'image-upload/style.scss';
 @import 'list/style.scss';
-@import 'list/collapsible-list/style.scss';
 @import 'order-status/style.scss';
 @import 'pagination/style.scss';
 @import 'pill/style.scss';
