+++ packages/experimental/src/experimental-list/collapsible-list/style.scss
@@ -1,4 +1,4 @@
-.woocommerce-list__item.list-item-collapse {
+.woocommerce-experimental-list__item.list-item-collapse {
 	justify-content: space-between;
 
 	p {
