+++ src/Loader.php
@@ -431,6 +431,14 @@ class Loader {
 		);
 		wp_style_add_data( 'wc-customer-effort-score', 'rtl', 'replace' );
 
+		wp_register_style(
+			'wc-experimental',
+			self::get_url( 'experimental/style', 'css' ),
+			array(),
+			$css_file_version
+		);
+		wp_style_add_data( 'wc-experimental', 'rtl', 'replace' );
+
 		wp_localize_script(
 			WC_ADMIN_APP,
 			'wcAdminAssets',
@@ -446,7 +454,7 @@ class Loader {
 		wp_register_style(
 			WC_ADMIN_APP,
 			self::get_url( "app/style{$rtl}", 'css' ),
-			array( 'wc-components', 'wc-customer-effort-score', 'wp-components' ),
+			array( 'wc-components', 'wc-customer-effort-score', 'wp-components', 'wc-experimental' ),
 			$css_file_version
 		);
